requirements.txt
# server
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0

# common
python-dotenv==1.0.1
pandas==2.1.2

# vision
opencv-python==4.8.1.78
python-multipart==0.0.9
ultralytics==8.0.203
dill==0.3.7

# LLM
langchain-openai==0.0.6
langchain_experimental==0.0.52
openai==1.12.0
langchain==0.1.8

# background
grobid-client==0.8.5
faiss-cpu==1.7.4
tiktoken==0.6.0
tabulate==0.9.0

